tracked down spec nonconformity, annotated source
During testing, we get a ton of spec output
yelling about non-conformity.  This comes from
the usage of
com.tbaldridge.odin.contexts.spec/spec-forms
which appears to scan the spec registry and
check for conformance to odin's internal
::spec definition.  Failure to conform
results in a printout - which ends up
being somewhat voluminous and induces
fears of regression.  In practice, it
looks like the print outs are harmless,
applying to the core clojure specs,
and not impacting test results.